Adds Lightning memo handling and received message logic
Introduces functions to process and format Lightning payment memos, including detecting and handling received payment messages.

Refines memo processing logic to ensure cleaner outputs and enhances memo-related functionality within transfer operations. Updates type annotations for better precision with monetary values.

Improves integration of newly added features across transfer and invoice processing workflows to streamline memo handling.

Fixes minor typographical error in docstring.

+++ src/v4vapp_backend_v2/hive_models/op_transfer.py
@@ -111,7 +111,7 @@ class TransferBase(OpBase):
             self.d_memo = self.memo
 
     @property
-    def amount_decimal(self) -> float:
+    def amount_decimal(self) -> Decimal:
         """Convert string amount to decimal with proper precision"""
         return self.amount.amount_decimal
